synchronize behavior between diff and hardlink
